10.1. Loading Animations & Quantity Input
Our add to cart button works... though it's not very obvious. We click "Add to Cart" and... nothing happens! Sure, we can refresh and see that the shopping cart header is increasing... but come on! We need some fireworks when an item is added!

Loading Animation
Find the data() section of the component. Our first mission is to add a loading animation while the item is being added. To track that "state", create a new addToCartLoading data set to false by default.
Then, inside of the method, put this.addToCartLoading = true on top... and this.addToCartLoading = false on the bottom.
Oh, but now we will need to make the AJAX call wait until it finishes. Add await and then, of course, we need to make the function async.
I love it! Copy the new addToCartLoading and head up to the template. Let's see.. right after the button, add an <i /> tag that's self-closing with v-show="addToCartLoading". Our project uses FontAwesome, so we can turn this into a loading animation with class="fas fa-spinner fa-spin".
Yes, we do have a Loading component, but this will look nicer in this spot. Back at the browser, we shouldn't even need to refresh. Yep! It instantly works and looks better.