10.2. Loading Animations & Quantity Input
Add to Cart Success State
But, hmm. I also want a user to get a happy feeling once the "add to cart" finishes. Back in the component, go find data again: I want to be able to track whether or not an "add to cart" has finished successfully. Create a new piece of data for this called addToCartSuccess set to false.
Now, in the method, after the AJAX call finishes, say this.addToCartSuccess = true. Oh, and in case we add multiple things in the cart, when the method starts, make sure this is false. So it starts false, then goes to true once we're done.
Back up in the template, we're going to use this in a very similar way. Copy the <i> tag, use addToCartSuccess and, for the class, use fas fa-check for a cool, not-spinning, check mark.
Let's try it! Back on the page, click and... oh. That felt good. I feel this uncontrollable need to keep shopping...