10.3. Loading Animations & Quantity Input
Hooking up the Quantity Input
Ok: when we call addItemToCart, in addition to product, there are two other pieces of data: color - for products that come in multiple colors - and quantity. Right now, both are hardcoded to boring values.
Let's fix quantity first. What we need to do is basically bind this input to a piece of data so that when the input changes, the data changes! And... we know how to do that!
Start by adding the new data key: quantity with a default value of 1.
Next, up in the template, Vue makes it beautifully simple: v-model="quantity".
That cool trick will convert the string input value into a Number type. Now we can use this below: instead of quantity: 1, we want quantity: this.quantity.
Thanks Vue! Now, when the input changes, the quantity data will change. And hmm, to be a bit fancier, I'll say v-model.number.
Head back to the browser. Nice! A moment ago, the input was blank. Now, it starts at 1. Increase this to 3. We currently have 12 items in the cart. Hit "Add to Cart", refresh and... woo! 15!
Ok: the last missing piece is the product color. Well, this product doesn't come in multiple colors, so it's finished. But if you click "Furniture" and then the inflatable sofa, this does come in multiple colors. Next: we need to grab the selected color and send that in the Ajax call. And you know what else? Whenever we hit "Add to Cart", I'm tired of the header not updating until we refresh. Let's also fix that... which is a bit interesting because it lives outside of our Vue component.