11.2. External DOM Updates
Updating the Shopping Cart Header Count
At this point. I'm pretty happy with our "add to cart" feature. Well, happy except for one detail: I don't like that the shopping cart count in the header doesn't update until I refresh the page.
But... what can we do? That isn't inside of our Vue app!
The answer is... who cares? What I mean is, if it's important for us to update this count for a good user experience, we can totally do it using good, boring JavaScript.
Open the template that holds this: templates/base.html.twig. I'll search for "shopping cart". Here it is: line 44. To make it easy to find this span in JavaScript, add an id="js-shopping-cart-items".
Next, back in the component, after we successfully add the item to the cart, we can say document.getElementById() - paste js-shopping-cart-items - then .innerHTML = getCartTotalItems().
That's a new function that we haven't used yet. When I hit tab to auto-complete it, PhpStorm added the new import for us... though I'm going to move this all back onto one line.
Anyways, the function comes from cart-service - the file we copied into our project a few minutes ago. Here it is. Very simply, it loops through the items and counts all of them using their quantity. A simple helper to get the number.
 149 lines  assets/services/cart-service.js
/**
 * Gets the total number of items in our shopping cart
 *
 * @param {CartCollection} cart
 * @return {number}
 */
export function getCartTotalItems(cart) {
    return cart.items.reduce((acc, item) => (acc + item.quantity), 0);
}
Back in product-show.vue, down in the method, we can call getCartTotalItems() and pass it this.cart. Because this returns a Number, call .toString() on the result.
So... no. This is not the most hipster code that you will ever write. But it will work and give us the great user experience we want. If I needed to update this header from multiple places in my code, I would definitely isolate it into its own JavaScript module to avoid duplication.
Let's see if it works! Back at the browser, make sure to refresh so the span gets the new id. Let's add 3 more green sofas. Watch the header... boom! That is so nice.
Next: with the add to cart done, let's create a new page: one that will display the shopping cart and checkout!