12.1. New Page, New Webpack Entry
Yay! Our customers can now add these top-quality products to their cart. Now we need a way for them to check out. Here's the plan: we're going to create a brand new page that holds a, sort of, "shopping cart" / "checkout" combo.
To create this page, if we want, we could build it entirely on the server-side with Twig. That's a great option if we don't need the power of Vue... which we definitely do not always need. But... you know, since this is a Vue tutorial, I was thinking that we should use Vue.

New Page Controller & Template
The first step to creating a page is... exactly the same whether it will use Vue or not: create a route, controller and template in Symfony. And... I've done that for us! Open src/Controller/CartController.php:

 20 lines  src/Controller/CartController.php
class CartController extends AbstractController
{
    /**
     * @Route("/cart", name="app_cart")
     */
    public function shoppingCart(): Response
    {
        return $this->render('cart/cart.html.twig');
    }
}

This shoppingCart() action will be called when we go to /cart and it renders a template: templates/cart/cart.html.twig
6 lines  templates/cart/cart.html.twig
{% extends 'base.html.twig' %}
{% block body %}
{% endblock %}
This holds a giant TODO!
The name of the route for this page is app_cart. So, as a first step, let's link to this from the shopping cart total at the top of the page. Open templates/base.html.twig. And... set the cart link href="" to {{ path('app_cart') }}
Go back to the browser... refresh and click the shopping cart. Hello big, gigantic <h1> tag!