14.1. More Mixin Magic
We created this mixin with the goal of sharing Vue logic between the product-show component and the new shopping-cart component that we're about to really start building. Specifically, both need a cart data and both need to make an Ajax call to populate that cart data. That stuff is now inside the mixin. Woo!
Both pages will also need the ability to add an item to the cart. It might seem weird to add an item to the cart from the shopping cart page, but we're going to add a "featured product" sidebar on this page with an "add the cart" button. So, in preparation for that, let's also move the add to cart functionality from product-show into the mixin.

Moving addToCart to the Mixin
This literally means that we're going to move the addToCart method from the component into the mixin. So let's copy it, delete it, and then, in the mixin, I'll add methods and then paste.
Oh, and this references two pieces of data that are related to this process: addToCartLoading and addToCartSuccess. Those should also now live in the mixin. Go find them, copy them, delete them, and paste them inside the mixin.
export default {
    data() {
        return {
            cart: null,
            addToCartLoading: false,
            addToCartSuccess: false,
        };
    },
};
If we stopped now, this probably would work because the new data and the new method are, as I mentioned, basically copied into our component. But it is a bit weird because the mixin references things like this.product and this.selectedColorId... but those don't live in the mixin. That's technically okay... and we sometimes do things like this in PHP with traits... which are very similar to mixins.
But I don't like. It makes the mixin hard to use correctly. The only way you can use this mixin is if your component has data with these exact names... which you would need to, kind of, dig a little to find. It also makes it easy to break your mixin... because if you renamed a piece of data in the component and updated all the references in that component... it wouldn't be very obvious that the mixin is also relying on this.