15.1. The Cart Page
Thanks to the mixin, our shopping-cart component has the cart data, which is being loaded via AJAX on created. Let's use that to build this page!
Start with a v-if so that we don't try to use the cart data before it's loaded. <div v-if= then cart !== null.
To know what we can do inside of this, go check out the Vue dev tools. The ShoppingCart component's cart data has an items key that we can loop over. Now, each item does not have any unique "id" property, which is a bit of an odd setup in our API, but it's fine. A unique id would really be the combination of the product IRI and the color IRI.
Anyways, let's loop over this: add div, put it on multiple lines, with v-for. I'm going to use the longer syntax so we can access the index: cartItem, index in cart.items.
I'm doing this because the cartItem doesn't have a unique id. So for right now, set :key to index. We'll improve that later with a true unique key, but it will work fine to start.
Inside the div, let's print cartItem.product and also cartItem.quantity.
Let's see how this looks! Move over to your browser. Hey! It works! It's not very pretty or useful yet... but it's a start! Crawl before you walk, as I like to say.

When the Cart is Empty
Before I forget, let's also handle the situation where the cart is empty... we want to give those customers some encouragement to keep shopping. Down below, but still inside the v-if for cart, add <div v-if="cart.items.length === 0", then we know that

  Your cart is empty. Get to shopping!
We can test this by messing with the Vue dev tools. Find the cart data, edit the items key and set it to an empty array. There we go!

Loading Animation
As a final touch, since the cart doesn't render until the cart AJAX call has finished, let's add a loading animation. Down on the component, start by importing our re-usable loading component: import Loading from '@/components/loading. Then add that to the components option... and in the template, use it with <loading v-show="cart === null" />.
We could also use v-if since the loading animation will be shown once and then hidden forever... but as we talked about, these are micro-optimizations. It doesn't really matter.
Oh, and later we might need a smarter loading mechanism that uses some loading data. But right now, checking the cart is perfect: if it's set, we know that we're ready to render!
Let's try it! Move over, refresh and... yep! It was quick - but we have a loading animation!