15.2. The Cart Page
We need More Data!
What's next? Well, we really need to render more data, like the product name and price. The problem is... we don't have that data! Look at the Vue dev tools and find the cart data. Each cart item has just these three keys. In a more perfect world, the cart API call might return more, like the actual product data instead of just the IRI string.
But since it doesn't, we're going to need to make another AJAX call to get the product data for each item in the cart.
Start by opening assets/services/product-service.js. At the bottom, I'm going to paste in a new function: fetchProductsById(), which you can copy from the code block on this page.
This is going to be really useful. It will allow us to collect the product IRI string for each item in the cart, then call this function to make one AJAX call to fetch all that product data at once.
But... how exactly can we do that? In shopping-cart, we need to call this new function after the cart AJAX call has finished... because we need that data to know which product IRIs to fetch. But the cart AJAX call doesn't live in this component: it lives over in created in the mixin. How can we run some code after this finishes?
Next, let's talk about how to do this. Then we'll get to work fetching and filling in all of our missing data thanks to a clever computed property.