16.1. Loading the "Complete" Cart
To render real product details on the cart page, we need more data than the original cart API returns.
So here is what we need to do: wait for the cart AJAX call to finish, collect the product IRI strings for each cart item, then make a second AJAX request for those product details. Normally, waiting for one AJAX call to finish before starting another one is easy. But in this case, the cart AJAX call lives in the mixin, in created()... and this new code needs to live in our shopping-cart component.
So: how can we run code after the cart AJAX call finishes? The answer is... a watcher. Remember: a watcher is a way to run a function whenever a piece of data changes. We rarely use them, because there are usually other ways to accomplish what we need, like listening to an event. But in this case, a watcher that watches for the cart data to change - specifically for it to change from null to an object - is probably our only simple option.

Adding the Watcher
Add a new watch key with a function called cart() inside... so that it's executed when the cart data changes. The first time this function will be called - and the only time in our app - is when the cart data change from null to an object.
Inside, we need to get an array of all the product IRI's for all the items in the cart. Do that with const productIds =, this.cart.items.map(). Pass a callback with an item argument. I'll use the very short syntax => item.product.
So: this will loop over all the cart items, call this function for each item, and we return the IRI string via item.product. The result is that productIds will be an array of the IRI strings: exactly what we want.
We can use that with the new function we copied a few minutes ago: const productsResponse = await - then fetchProductsById(). Let PhpStorm autocomplete that so it adds the import. Pass this productIds. And, of course, as soon as we add await, the function needs to be async.
To see what this looks like, let's console.log(productsResponse).
Cool! Move over, refresh, go to the console and.... nice! The response has the usual data key... with a hydra:member property that holds an array of all those beautiful products.
Oh, and to help us understand the next step, let's also log this.cart. Remember: the cart object has an items key and each item has color, product and quantity. The problem is that product and color are just IRI strings, instead of real data.
Because of that, looping over this.cart.items inside our template is not very useful. Instead, I want to create a new array of cart items that has the same structure - color, product and quantity keys - but where product and cart are set to data objects, instead of strings. Then, we can loop over that in the template and have everything we need. What could go wrong?