16.2. Loading the "Complete" Cart
Creating the Complete Cart Data
Back in our component, remove the log and say const products = productResponse.data['hydra:member'] to get to where the actual data is stored.
Now create a new object: const completeItems. This is going to be kind of cool: set it to this.cart.items.map(). So once again, this will loop over each item and call our function. Give it the cartItem argument, and, just like last time, I'm going to use the short syntax... but since I want to use multiple lines, I'll add parenthesis.
Inside, return an object, well... thanks to the short syntax, the return statement is implied - our function will return this object. And this object will have the same keys as the cart items, like product! But this time, we want to set it to the product object. And we can use the products array to find it: products.find(), pass an arrow function with a product argument, and return true if product['@id'] equals cartItem.product. So, we're comparing the IRI strings.
You could choose to write some extra code in case an item in the cart is not returned from the products AJAX call, like maybe it was removed from our system. But I'll skip that.
Next, set color to cartItem.color. That's still the IRI string... not a color object - but we'll fix that soon. And then quantity is simple: set that to cartItem.quantity.
To celebrate, let's see what this looks like! console.log(completeItems).
Back at the browser... I don't even need to refresh: the log is already here: 3 items... which are much more useful. The color is still a string, but product is an object.
If we loop over this, we're going to be a lot more dangerous.
So the question now is: how can we make this new completeItems array accessible to our template? The easiest thing to do would be to set it as a new key on data. But... that would not be our best idea. Let's talk about why next and find a solution that, honestly, I love.