18.1. Parallel AJAX with Promises
The last missing piece of data on the cart page is the item color. I want to show a box with the actual color... but for that, we need the hex value. All we have right now, if we look at the Vue Dev tools, is the color IRI. Of course, with another AJAX call, we could use that to get the color data!
Cool! Let's go! We'll do exactly what we did to fetch the product data. First, add a new colors data set to null
Then, below in the watcher, we can fetch all the colors and then set them on that data. Do that with const colorsResponse = await and then call a function we saw earlier: fetchColors(). Hit tab to auto-complete that so PhpStorm adds the import way up on top
Back in the watcher function, all we need is this.colors = colorsResponse.data['hydra:member']
We could make this smarter - like we did with products - where we only fetch the color data that we need based on the items in the cart. But for our site, there are very few total colors, so I've decided to fetch all of them.
Now update completeCart(). First, if the colors data is not set yet - if not this.colors - then we should also return null from completeCart().
Inside the object, use the same trick we used to find the matching product: paste, and change the line to this.colors, color, color, and cartItem.color.
Inside the object, use the same trick we used to find the matching product: paste, and change the line to this.colors, color, color, and cartItem.color.
Testing time! Back at the browser... woohoo! My two inflatable sofas have a color. Those are going to look great in the office. Not all products have a color, but those that do, are now printing it.